storing rows as absolutes

+++ src/runtime_src/xdp/profile/plugin/aie_trace_new/aie_trace_metadata.h
@@ -63,7 +63,7 @@ class AieTraceMetadata{
     std::string getMetricSet(const std::string& metricsStr);
 
     int getHardwareGen();
-    int getAIETileRowOffset();
+    uint16_t getAIETileRowOffset();
     std::vector<tile_type> getTilesForTracing();
     std::vector<tile_type> getMemTilesForTracing();
